FROST-8496 - ABP_Player_v2.uasset: [Compiler] Can't connect pins: float is not compatible with Double Real.
A couple of BP reals changes were missed for IKRig and a new-ish function in RigVMTypeUtils.

For RigVMTypeUtils, we simply need to ensure that we're using the correct PC_Real category when dealing with float/double CPPTypes.

For IKRig, the alpha pin needs to be a real/double type. Additionally, when copying the property data to the FIKRigGoal instance, we need to treat the property's value as a double type. Since FIKRigGoal::Position et al. use the alpha as a normalized value, it likely doesn't need the precision of a double, so it's been left as is. As a result, we need to perform a narrowing conversion when copying over the data.

#jira FROST-8496
#preflight 61faf7115e35b9215b000da8
#rb julien.marchand
#lockdown julien.marchand

[CL 18835074 by dave jones2 in 5.0 branch]