UE-141088 - Crash occurs when reading Real values from None objects
By default, implicit cast instructions expect their source expression to evaluate correctly and return a non-null source property to cast. If we read a variable from a context that's null (eg: a null actor's field), then the expression is never evaluated, and a null property is set as the most recent property, which trips an internal check.

While the attempted null access on the context triggers an internal Blueprint exception, we don't have any stack unwinding behavior. As a result, it continues evaluation. Our only protection is to check the most recent property's validity after stepping through the stack and before performing the cast.

Our solution uses the new StepAndCheckRecentProperties convenience function to check the state of the most recent property. If null, we simply skip the cast and log the error.

#rb phillip.kavan
#jira UE-141088
#preflight 61fc432c033a864b77fd33d0
#lockdown julien.marchand

[CL 18854053 by dave jones2 in 5.0 branch]