Fix crash when using the GPU benchmark function.
The code was trying to create a readback-enabled render target, which is no longer supported now that pooled render targets no longer distinguish between targetable and shader resources. The target wasn't needed anyway, it was used in some code which was trying to sync the GPU with a readback that's been commented out 7 years ago. This CL removes the target entirely.

#jira UE-141268
#rnx
#lockdown Michal.Valient
#preflight 620100856f05911039b3733a
#rb Tiago.Costa

[CL 18884247 by mihnea balta in 5.0 branch]