nDisplay: Fix crash when renaming a viewport or cluster node to a previous used name. This was caused from engine changes where removing an instanced map element will rename the object to the transient package for us, causing a naming conflict.
Also fixed an issue where nDisplay's unique name creation wouldn't check the object outer. Not doing so could cause a crash when undoing the creation or rename of a node, then trying to assign another node the name of the node that no longer exists. That object would still be parented to the old outer but marked pending kill, despite proper transaction setup.

#jira UE-141494
#preflight 6201510c6773a3612899bdc3
#rb alejandro.arango, trystan.binkley-Jone
#lockdown alejandro.arango

[CL 18892493 by Patrick Hardy in 5.0 branch]