#ue5 - Fix Characters using PerchRadiusThreshold to not stand on the edge of geometry using the wrong location for the perch check.
Using the incoming Hit.Location as the capsule location was incorrect when the previous tests shrank the capsule height to avoid penetrations, placing the origination of the perch check often starting in penetration and causing a fallback to another shrunken test, which would waste performance and sometimes even fail. This was made more common when using bUseFlatBaseForFloorChecks, as the tests were often rejected when physics returns points on the edge of the box which was too close to the radius to be acceptable. In the worst case, no floor checks would pass when using bUseFlatBaseForFloorChecks (normal tests would be too close to edge, perch check would be in penetration).

#jira UE-141031
#rb Justin.Hare
#lockdown Julien.Marchand
#preflight 620435f38e53010a1872a301

[CL 18928239 by Zak Middleton in 5.0 branch]