Fix path tracer not being able to be toggled through the show command in game viewports
The code was improperly trying to change the ViewModeIndex which had no way of being reset afterwards. It turns out this logic had no effect and could be safely removed. Setting ShowFlags.PathTracing is sufficient to enable the path tracer, the ViewModeIndex is not actually used.

This enables a mechanism for toggling the path tracer in shipping builds.

#jira UE-143246
#preflight 62104d34ff52bfecfc0893d2
#rb Juan.Canada
#lockdown Juan.Canada

[CL 19086947 by chris kulla in 5.0 branch]