UE-145269 - Math Expression nodes failing within functions
We were only considering the first linked pin when iterating over the connections for the input pins. An input pin can be connected to multiple internal nodes, so we need to process those. For example, this problem would manifest if the expression were (x+x), and we connected the x input to a single precision float.

#jira UE-145269
#preflight 622a6bfe031cac3623f01924
#rb Ben.Hoffman
#lockdown

[CL 19342336 by dave jones2 in 5.0 branch]