DDC: Fixed incorrect cache propagation behavior
Fixed duplicate puts caused by the in-flight memory cache. Without using StopGetStore, get hits on the in-flight cache would propagate as puts to subsequent nodes in the hierarchy. Since the in-flight cache is only used as temporary storage while a put is in flight, this caused duplicates of the in-flight puts, which manifests as move collision log messages from the file system cache store.

Fixed incorrect skipping of get propagation past nodes with the flag StopStore. The wrong response status was being checked, and caused subsequent nodes to be skipped even if the StopStore node did not contain the data, if any previous node contained the data.

#jira UE-149602
#lockdown Mark.Lintott
#preflight 625eece03e0f6f80adb79d99
#rb Zousar.Shaker
#ushell-cherrypick of 19756375 by Devin.Doucette
#ushell-cherrypick of 19757142 by Devin.Doucette
#ushell-cherrypick of 19763306 by Devin.Doucette
#ushell-cherrypick of 19785647 by Devin.Doucette

[CL 19833748 by Devin Doucette in 5.0 branch]