OpenXR: properly initialize pipelined XrSpaceLocation
We were using SetNum for our DeviceLocations TArray. This lead to garbage data in DeviceLocations, which would accidentally populate LocationFlags.
Because we rely on caching for 'last known' locations, this leads to bugs when we fetch a partially valid location.
One specific bug scenario: we fetch a valid position, but invalid orientation. We reset position correctly in the cached locations. But the garbage data
is left for orientation, leading to failures in GetCurrentPose.

#jira UE-148686
#rb Jeff.Fisher
#preflight 6261a9c7d929bc34a3ee1967

[CL 19856731 by robert srinivasiah in 5.0 branch]