#jira UE-144710: Datasmith-generated UV scale is wrong The algorithm of UV Scaling is reimplemented based on TechSoft SDK + BReps are converted in CAD modeler (CADKernel, TechSoft, ...) in mm + improve naming
#jira UE-145444 : Retessellate "Skip Deleted Surface" does not work with TechSoft
PatchId wasn't set at the import

#jira UE-144502: Materials are not imported with TechSoft
In the jira case, the material is defined at the part level but not at the BRepModel level. The material of the part is assigned to the Body
Fix TechSoft Color definition

#jira UE-143597: Errors when importing a specific Solidworks file with sew option
the error appears with empty mesh (mesh without triangles). To avoid the error, meshActors with empty mesh are not build

#jira UE-147856: Datasmith incorrectly imports 3Dxml files
Error in AddFaceTriangleStripeWithUniqueNormal method (and also in AddFaceTriangleStripeWithUniqueNormalAndTexture method)

#Jira UE-149181
Fix scale on Meshing parameters. The scale was apply two times

#rb jeanluc.corenthin
#swarm
#preflight https://horde.devtools.epicgames.com/job/6266bf3e5037f70ba0eac4b8

#robomerge deadend

[CL 19908519 by David Lesage in 5.0 branch]