UE-149122 - TMap Values in User Structs do not convert to real numbers
The underlying cause is connected to inconsistent usage of serialization methods among BP-related types.

Both FEdGraphPinType and FStructVariableDescription have a FEdGraphTerminalType field type. FEdGraphTerminalType will only perform versioning fixup in operator<<. This implies that fixup can only be achieved when an outer type uses manual serialization and calls << on the type, which FEdGraphPinType does. However, FStructVariableDescription does *not* have a Serialize function, and expects tagged property serialization to fixup everything in its FEdGraphTerminalType member, which never happens.

Since this is needed as a hotfix, the easiest fix is to update FStructVariableDescription::PostSerialize to perform the fixup for FEdGraphTerminalType, which is what FEdGraphTerminalType already does in operator<<.

#jira UE-149122
#preflight 6271561f9d6c2f8f5b234a34
#rb marc.audy

[CL 20027908 by dave jones2 in 5.0 branch]