UE-148560 - Unable to bind Float variables to certain properties in Widget elements
Since BP floats default to double precision, we need to relax binding requirements. Most UMG widgets uses delegates that expect single-precision floats, so we need to allow float<->double conversion in the UFloatBinding layer. Without this change, we can't bind any new float variables in UMG, which makes the feature fairly useless.

The fix simply updates UFloatBinding::IsSupportedSource to allow double properties. Additionally, when we need to retrieve the underlying value, we need to inspect the property first. If it's a double, we'll perform a cast. Otherwise, it defaults to existing behavior.

#rb patrick.boutot
#preflight 6262f238cf17922036d8d169
#jira UE-148560

[CL 20380365 by dave jones2 in 5.0 branch]