Fix for HDR toggling on some platforms.
This does not truly change HDR since these platforms do not support toggling this at runtime. Instead we change the swapchain to be in SDR, and we end up in SDR mode with HDR output mode.

SDR in HDR mode is slightly darker on the TV, more than when comparing the screenshots of the swapchain. This is most likely due to the TV response, which needs to be investigated later on to see if we can somehow match the 2 properly.

In addition to the 2 CLs from main, this CL allows BP to change the CVar r.HDR.EnableHDROutput . On these platforms, the CVar is set with ECVF_SetByCode as soon as HDR was supported and calling EnableHDRDisplayOutput with ECVF_SetByGameSetting would get ignored

#jira UE-148923
#preflight 628fbfbd4f63120d8ee7af3d

Original CLs:
19334984
fix for r.setres on d3d12: the backbuffers were still referenced by the present queue, preventing their proper destruction

19609874
support for HDR change on process resume

[CL 20439865 by benjamin rouveyrol in 5.0 branch]